Update API renames/usage

+++ rmf_site_editor/src/site/model.rs
@@ -699,7 +699,7 @@ pub fn make_models_selectable(
     // If layer should not be visible, don't make it selectable
     if scene_roots
         .get(req.parent)
-        .is_ok_and(|r| r.iter().all(|l| l == MODEL_PREVIEW_LAYER as usize))
+        .is_ok_and(|r| r.iter().all(|l| l == MODEL_PREVIEW_LAYER))
     {
         return req;
     }
