+++ rmf_site_editor/src/site/primitive_shape.rs
@@ -85,10 +85,7 @@ pub fn handle_new_primitive_shapes(
             .last()
         {
             entity_commands.insert(render_layer.clone());
-            if !render_layer
-                .iter()
-                .all(|l| l == MODEL_PREVIEW_LAYER as usize)
-            {
+            if !render_layer.iter().all(|l| l == MODEL_PREVIEW_LAYER) {
                 spawn_selectable(entity_commands);
             }
         } else {
